x3: Remove oplus logical partitions when necessary
Usually when building full OTAs for dynamic partition devices the
whole partition table is removed and recreated, but in vendorless
cases lineageos (and basically any other custom rom) has a patch to
disable that functionality and instead just resize the partitions.
This, however, breaks the flashing if a new partition is added, because
that partition is tried to be resized and, since it doesn't exist, the
assert fails. This patch adds another dynamic patch list to remove few
useless Oplus partitions or gracefully fail if it already removed (no assert).

Signed-off-by: Kuba Wojciechowski <nullbytepl@gmail.com>
Signed-off-by: TheMalachite <eliasgheeraert@gmail.com>
Signed-off-by: Sukrut4778 <sukrutbhagwat4778@gmail.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>